requirements.txt
py7zr==0.21.0
PyYAML==6.0.1
rarfile==4.2